refactor: 가진 패의 점수 계산 책임 Hand 클래스에 이동
