+++ src/main/java/domain/Hand.java
@@ -1,5 +1,6 @@
 package domain;
 
+import constant.GameConstant;
 import domain.dto.CardDto;
 
 import java.util.ArrayList;
@@ -7,6 +8,9 @@ import java.util.Iterator;
 import java.util.List;
 
 public class Hand implements Iterable<Card> {
+    private static final int ACE_BONUS_SCORE = 10;
+    private static final int ACE_NO_BONUS = 0;
+
     private final List<Card> cards = new ArrayList<>();
 
     public void add(Card card) {
@@ -24,6 +28,29 @@ public class Hand implements Iterable<Card> {
     public boolean hasAce() {
         return cards.stream()
                 .anyMatch(c -> c.getCardRank().equals(CardRank.ACE));
+    }
+
+    public int calculateScore() {
+        int total = 0;
+        for (Card card : cards) {
+            total += card.getCardRank().getValue();
+        }
+        return total;
+    }
+
+    private int calculateAceScore() {
+        if (!hasAce() || calculateScore() > 11) {
+            return ACE_NO_BONUS;
+        }
+        return ACE_BONUS_SCORE;
+    }
+
+    public int calculateFinalScore() {
+        return calculateScore() + calculateAceScore();
+    }
+
+    public boolean isBust() {
+        return calculateFinalScore() > GameConstant.BUST_THRESHOLD;
     }
 
     public List<CardDto> toCardDtos() {
